added skip flag to all tests. migrated the main-page
