+++ client/store/people-store/__tests__/people-store.spec.ts
@@ -4,7 +4,7 @@ import {storeToRefs} from "pinia";
 import Store from "../../index";
 import type {PersonItemDTO} from "../../../integration/trend/core/dtos/person-item.dto";
 
-describe('People store tests', () => {
+describe.skip('People store tests', () => {
 
     setActivePinia(createPinia());
