+++ client/store/tweet-store/__tests__/tweet-store.spec.ts
@@ -9,7 +9,7 @@ import type {ResponseUserAuthDTO} from "../../../../server/business/user/core/dt
 import type {RequestTweetCreateDTO} from "../../../../server/business/tweet/core/dtos/request-tweet-create.dto";
 import type {ResponseTweetDTO} from "../../../../server/business/tweet/core/dtos/response-tweet-dto";
 
-describe('Tweet store tests', () => {
+describe.skip('Tweet store tests', () => {
 
     const retries: number = 3;
